Variable numeric: fix PHP 8.2 compatibility #743322
As well as the typical declaring of previously undeclared
fields, and cleaning up when fields that are not required
are being set, this change also sorts out the loading of
units in varnumunit.

The way it is supposed to work is that:
* get_question_options loads all the required data.
  (This was being done.)
* Then initialise_question_instance initialises all the required
  data on the ..._question object. This was missing.
Not doing this breaks the effectiveness of Moodle's caching of
question data.